backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
SQLAlchemy==1.4.53
PyMySQL==1.1.0
PyJWT==2.10.1
requests==2.31.0
paramiko==4.0.0
psutil==7.2.1
python-dotenv==1.0.0
watchdog==3.0.0
cryptography==41.0.4
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1
python-docx==1.2.0
PyPDF2==3.0.1
playwright==1.58.0
geoip2==4.7.0
